Fix comment inconsistency for ResizeUsePciRbIo quirk
Fixed a issue where O3C was not properly commenting on the ResizeUsePciRbIo quirk in the user's config.plist file. Now, if the quirk is disabled, a comment will be added explaining that it can improve compatibility and stability with certain hardware configurations but should only be used if necessary. If the quirk is enabled, no comment will be added assuming the user enabled it on purpose.