Refactor cpu_arch_list array for reusability in multiple files
The cpu_arch_list array was refactored to be a shared constant between
the plistParser.js and Reader.jsx files. This was done by defining the
array as a const variable in a separate file and importing it into both
files. In plistParser.js, only the value property of each object is used,
while in Reader.jsx both the value and text properties are used. This
change improves code maintainability by reducing duplication and making
it easier to update the list of supported CPU architectures.